Slowly page through Tesla vehicles
This won't affect many people. The goal is to have requests that are
under 10 seconds, which is the timeout on the Tesla proxy. We could
change that timeout but a 10 second request is absurd.

I suspect that it's the state checking that takes them a long time. I
wish we could skip it.